Refactor into `Simp::Packer::` classes
